\name{icCom}
\alias{icCom}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Compute AIC and BIC of VAR(1) model
}

\usage{
icCom(x,type,lambda,gamma,alpha)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
\item{type}{"aic", "bic".}

\item{...}
}

\seealso{
\code{\link{Breg}}
}
